pyproject.toml
[project]
name = "imageoptimiser"
version = "1.0.0"
description = "This script removes EXIF-data from images in the specified directory"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
]

[tool.ruff.format]
quote-style = "single"
